chore: separate out database files
