+++ packages/common/amplify_db_common/example/bin/database.dart
@@ -1,0 +1,41 @@
+// Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// The example below is taken from https://drift.simonbinder.eu/docs/getting-started/
+
+import 'package:amplify_db_common/amplify_db_common.dart';
+import 'package:drift/drift.dart';
+
+part 'database.g.dart';
+
+class Todos extends Table {
+  IntColumn get id => integer().autoIncrement()();
+  TextColumn get title => text().withLength(min: 6, max: 32)();
+  TextColumn get content => text().named('body')();
+  IntColumn get category => integer().nullable()();
+}
+
+@DataClassName('Category')
+class Categories extends Table {
+  IntColumn get id => integer().autoIncrement()();
+  TextColumn get description => text()();
+}
+
+@DriftDatabase(tables: [Todos, Categories])
+class MyDatabase extends _$MyDatabase {
+  MyDatabase() : super(connect(name: 'MyDatabase', path: '/tmp'));
+
+  @override
+  int get schemaVersion => 1;
+}
