+++ packages/common/amplify_db_common/example/web/main.dart
@@ -12,8 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import 'package:amplify_db_common_example/database.dart';
 import 'package:example_common/example_common.dart';
+
+import 'database.dart';
 
 final db = MyDatabase();
