+++ packages/common/amplify_db_common/test/main_test.dart
@@ -18,8 +18,7 @@
 import 'package:amplify_db_common/amplify_db_common.dart';
 import 'package:test/test.dart';
 
-// ignore: avoid_relative_lib_imports
-import '../example/lib/database.dart';
+import 'database.dart';
 import 'util.dart';
 
 void main() {
